updating code for better clarity, renamed FilterHelper to RequestUtils, moved authentication code to new bean JwtAuthenticationHelper
